Fix and Changes
Removed Provider, Requester in favor of just being a user field in relative parts
Add User info on getbyid of model and requests.
Changes to how Create works, There is a createmodel now that will take the needed information for creation. No longer do we need everything of User only the ids. This could cause problems later with security but for now its going the right direction I believe.
Changes to seed data to connect to relative subdata. This way we can actually see data when testing.
Might have changes that i forgot to note. Prob isn't that important...